New: Expand parent mechanism stereochemistry
Expands parent mechanism stereochemistry based on a submechanism
stereoexpansion.

The expansion is naive and simply "unlumps" the reactions of expanded
species in a way that reproduces the original model (assumes an equal
ratio of stereoisomers concentrations).

To arrive at a proper treatment of stereochemistry for the expanded
species, the resulting mechanism should be merged with the submechanism,
dropping all reactions from the parent mechanism that are treated in the
submechanism.